Doc: Improve documentation for QObject::findChild
There is a difference between findChild() and findChild("") that was
not explained in the documentation. So, add a sentence to clarify it.

The object's search when there are several direct ancestors was
documented as "undefined". Change it to "first child in children()".

Task-number: QTBUG-103986
Pick-to: 6.5
Change-Id: I25cb01b87d3b24af498ad494b3f353840ee78807
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
(cherry picked from commit 80cdd4e5bbe15ef3235b1269b3781f8e4b19a5fa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>